api/requirements.txt
Flask==2.3.2
Flask-SocketIO==5.3.4
kafka-python==2.0.2
six==1.16.0